+++ scripts/build.js
@@ -8,11 +8,12 @@ module.exports.build = async () => {
     const ctx = await utils.context();
 
     console.log(`Clearing template dir: ${ctx.testDir}`);
-    sh.rm("-r", ctx.testDir);
 
     if (!fs.existsSync(ctx.testDir)) {
-        sh.mkdir(ctx.testDir);
+        throw new TypeError(`Expected directory ${ctx.testDir}`);
     }
+
+    sh.rm("-r", path.join(ctx.testDir, "*"));
 
     console.log(`Deploying template to ${ctx.testDir}`);
     sh.cd(ctx.testDir);
@@ -29,9 +30,19 @@ module.exports.build = async () => {
     const templateDir = path.dirname(templates[0]);    
 
     // sh.cd won't work because it's not the current shell you execute this on.
-    console.log(`Project generated: ${templateDir}`);
+    console.info("Execute the following:");
+    console.info(`- cd ${templateDir}`);
+    console.info("- npm install");
+    console.info("- npm run build");
+    console.info("- npm run startSAM (for the AWS SAM api)");
+    console.info("- npm run startExpress (for the standalone express app)");
 };
 
-this.build().then(() => {
-    console.log("Done");
-});
+this.build()
+    .then(() => {
+        console.log("SUCCESS");
+    }
+    ).catch((err) => {
+        console.error("FAILURE");
+        console.error(err); 
+    });