+++ {{cookiecutter.name}}/api/src/lambda.ts
@@ -1,10 +1,10 @@
-'use strict'
-import { createServer, proxy, Response } from "aws-serverless-express";
+"use strict";
 import lambda from "aws-lambda";
+import { createServer, proxy, Response } from "aws-serverless-express";
 import { app } from "./app";
 
 const server = createServer(app);
 
 export const handler = (event: lambda.APIGatewayEvent, context: lambda.Context) => {
     return proxy(server, event, context);
-}
+};