+++ {{cookiecutter.name}}/scripts/startExpress.js
@@ -1,0 +1,19 @@
+const utils = require("./utils");
+const find = require("find");
+const path = require("path");
+const sh = require("shelljs");
+
+module.exports.startApi = async () => {
+    const ctx = await utils.context();
+
+    utils.execInDir(ctx.rootDir, "node ./dist/app.local.js");
+};
+
+this.startApi()
+    .then(() => {
+        console.log("SUCCESS");
+    }
+    ).catch((err) => {
+        console.error("FAILURE");
+        console.error(err); 
+    });